start second part

+++ 42-cursus/libft/ft_strlen.c
@@ -18,7 +18,7 @@
 */
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
